requirements.txt
altgraph==0.17.2
astroid==2.7.3
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
paramiko==2.7.2
pkg_resources==0.0.0
platformdirs==2.3.0
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pylint==2.10.2
PyNaCl==1.4.0
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
wrapt==1.12.1
